spring-mvc/demo-web-mvc/src/main/java/me/rldnd/demowebmvc/SampleController.java
package me.rldnd.demowebmvc;

import org.springframework.http.MediaType;
import org.springframework.stereotype.Controller;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.ResponseBody;
import org.springframework.web.bind.annotation.RestController;

import java.awt.*;

@Controller
public class SampleController {

    @RequestMapping(
            value = "hello",
            consumes = MediaType.APPLICATION_JSON_VALUE,
            produces = MediaType.TEXT_PLAIN_VALUE
                )
    @ResponseBody
    public String hello() {
        return "hello";
    }


}
